Create CrashesByYearCumulative component for cumulative chart
